requirements.txt
# Chess logic and UCI engine (for optional Stockfish in terminal)
python-chess>=1.999

# Load .env (API keys, STOCKFISH_*)
python-dotenv>=1.0.0

# S3 state sync when STATE_BUCKET is set (backend game runner)
boto3>=1.34.0

# LLM API clients
openai>=1.0.0
google-genai>=1.0.0
anthropic>=0.40.0
mistralai>=1.0.0
cohere>=5.0.0
groq>=0.4.0

# Web UI (optional)
streamlit>=1.33.0

# Frontend API server (optional): uvicorn api:app
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
